+++ functional-test/src/main/java/org/zanata/page/languages/ContactTeamPage.java
@@ -31,18 +31,11 @@ import org.zanata.page.BasePage;
 @Slf4j
 public class ContactTeamPage extends BasePage {
 
-    private By subjectField = By.id("contactCoordinatorForm:subjectField:subject");
-    private By messageField = By.id("contactCoordinatorForm:messageField:message");
-    private By sendButton = By.id("contactCoordinatorForm:send");
+    private By messageField = By.id("contactCoordinatorForm:messageField:contact-coordinator-message");
+    private By sendButton = By.id("contact-coordinator-send-button");
 
     public ContactTeamPage(WebDriver driver) {
         super(driver);
-    }
-
-    public ContactTeamPage enterSubject(String subject) {
-        log.info("Enter subject {}", subject);
-        waitForWebElement(subjectField).sendKeys(subject);
-        return new ContactTeamPage(getDriver());
     }
 
     public ContactTeamPage enterMessage(String message) {