+++ zanata-war/src/main/java/org/zanata/adapter/PropertiesAdapter.java
@@ -1,0 +1,7 @@
+package org.zanata.adapter;
+
+/**
+ * @author Alex Eng <a href="mailto:aeng@redhat.com">aeng@redhat.com</a>
+ */
+public class PropertiesAdapter {
+}